add-forecast. add forecast
