+++ src/utils/interfaces/IMinutelyForecast.ts
@@ -1,0 +1,5 @@
+
+export interface IMinutelyForecast {
+    dt: number;
+    precipitation: number;
+}
