+++ src/utils/interfaces/IWeather.ts
@@ -1,0 +1,6 @@
+export interface IWeather {
+    id: number;
+    main: string;
+    description: string;
+    icon: string;
+}
